# Optimize cache flush to fix lock contention in concurrent d2i_X509 calls